+++ src/shell/commands/rmdir.ts
@@ -2,9 +2,9 @@ import { fsDB } from '@/db/fs';
 import { Executor } from '../executeCommand';
 import { partitionArgs } from '../utils';
 import { myPath } from '@/modules/myPath';
-import { deleteDirectory } from '@/db/systemCalls/deleteDirectory';
-import { resolveInodeId } from '@/db/systemCalls/utils/resolveInodeId';
-import { SysError } from '@/db/systemCalls/utils/SysError';
+import { deleteDirectory } from '@/systemCalls/deleteDirectory';
+import { resolveInodeId } from '@/systemCalls/utils/resolveInodeId';
+import { SysError } from '@/systemCalls/utils/SysError';
 
 const executor: Executor = async (argv, log) => {
   const { flags, positionals } = partitionArgs(argv);
