+++ src/systemCalls/deleteDirectory.ts
@@ -1,4 +1,4 @@
-import { FsDB } from '../fs';
+import { FsDB } from '@/db/fs';
 import { serializeDirectoryContent } from './utils/dir';
 import { readDirectory } from './readDirectory';
 import { SysError } from './utils/SysError';
