+++ src/systemCalls/readDirectory.ts
@@ -1,4 +1,4 @@
-import { fsDB, FsDB } from '../fs';
+import { fsDB, FsDB } from '@/db/fs';
 import { deserializeDirectoryContent } from './utils/dir';
 import { readDirectoryBlob } from './utils/readDirectoryBlob';
 import { resolveInodeId } from './utils/resolveInodeId';
