+++ src/systemCalls/utils/readDirectoryBlob.ts
